% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dodec.R
\name{Spc.MakeDodecahedron}
\alias{Spc.MakeDodecahedron}
\title{Return a dodecahedron  object, constructed from elementary spacial objects}
\usage{
Spc.MakeDodecahedron(centre, radius, properties = NA, bound = TRUE,
  align.to = "x")
}
\arguments{
\item{centre}{Centre of dodecahedron}

\item{radius}{Is the nominal size from the centre to each vertex. So the radius of the smallest sphere containing the object.}

\item{properties}{Package-independent object defining additional properties}

\item{bound}{Include a bounding sphere? Default=TRUE}

\item{align.to}{Name of axis to align to - i.e. 2 surfaces will be level in that plane. Default = "x"}
}
\value{
A compound object, comprising of elementary triangle objects
  in a grouped hierarchical object, with a surrounding bounding sphere
  used for spacial indexing.
}
\description{
Return a dodecahedron  object, constructed from elementary spacial objects
}
\examples{
  w <- Spc.MakeDodecahedron(c(0,0,0), 10, surface_props)
}
