% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/teapot.R
\name{Spc.MakeTeapotahedron}
\alias{Spc.MakeTeapotahedron}
\title{Return a teapotahedron  object, constructed from elementary spacial objects}
\usage{
Spc.MakeTeapotahedron(centre, radius, properties = NA, bound = TRUE,
  align.to = "x")
}
\arguments{
\item{centre}{Centre of teapotahedron}

\item{radius}{Is the nominal size from the centre to each vertex. So the radius of the smallest sphere containing the object.}

\item{properties}{Package-independent object defining additional properties}

\item{bound}{Include a bounding sphere? Default=TRUE}

\item{align.to}{Name of axis to align to - i.e. 2 surfaces will be level in that plane. Default = "x"}
}
\value{
A compound object, comprising of elementary triangle objects
  in a grouped hierarchical object, with a surrounding bounding sphere
  used for spacial indexing. It should be noted that this is not an exclosed solid, and
  that the surface normals need a bit of work
}
\description{
Return a teapotahedron  object, constructed from elementary spacial objects
}
\examples{
  w <- Spc.MakeTeapotahedron(c(0,0,0), 10, surface_props)
}
\seealso{
Other constructors: \code{\link{Spc.MakeDodecahedron}}
}
